Add GitHub sync webhook endpoint and logic to process GH issues webhooks
The received webhooks calls are validated against their signature using
the authors_github["webhook_key"] encrypted credential, which has to be
set up in GitHub as the webhook key.
The events processed are only the issue related actions "opened", "edited",
and "deleted"; and they accordingly create, update or delete authors in
the database.
The processing of the events is done asynchronously through an Active Job,
this allows both reprocessing errors, and scaling the event processing
separately from the API backend.

+++ app/models/github_sync_service.rb
@@ -6,6 +6,7 @@ class GithubSyncService
   end
 
   # Syncs the github issues with authors
+  #
   # - Creates issues for authors that don't have a github issue number in the model (and updates their github issue number)
   # - Updates existing authors with github issue numbers from the github issue info (only if there were modifications)
   # - Deletes authors with github issue numbers if the github issue doesn't exist anymore
@@ -20,6 +21,33 @@ class GithubSyncService
     Rails.logger.error "Aborting synchronization"
     Rails.logger.error "There was an error syncing with GitHub"
     raise err
+  end
+
+  # Syncs the database from a received github event.
+  #
+  # This method is idempotent, in the sense that an event can be processed multiple times and it won't have unintended
+  # side effects.
+  #
+  # Valid events are:
+  # - create: Creates a new author with book from the issue information. If there is already an author with that issue
+  #   number, it updates the existing author
+  # - update: Updates an existing author with book from the issue information. If the author doesn't exist create it
+  # - delete: Deletes an existing author with the issue number. If there are no authors with that issue number does nothing
+  def sync_from_event event
+    action = event["action"]
+    github_issue = event["issue"]
+    issue_number = github_issue["number"]
+    author = Author.find_by_gh_issue_number(issue_number)
+    case action
+    when "edited"
+      update_or_create_author_from_github(author, github_issue)
+    when "opened"
+      create_or_update_author_from_github(author, github_issue)
+    when "deleted"
+      delete_author(author, github_issue["number"])
+    else
+      Rails.logger.warn "Unhandled event action '#{action}' received from GitHub event"
+    end
   end
 
   private
@@ -38,10 +66,11 @@ class GithubSyncService
   def update_authors_synced_with_issue
     authors_to_update = authors_synced.select { |author| issue_numbers.include?(author.gh_issue_number) }
     authors_to_update.each do |author|
-      if author_is_up_to_date(author)
+      github_issue = github_issue_for author
+      if author_is_up_to_date(author, github_issue)
         Rails.logger.info "Author #{author.id} is up to date, no need to update it"
       else
-        update_author_from_github(author)
+        update_author_from_github(author, github_issue)
       end
     end
   end
@@ -57,8 +86,7 @@ class GithubSyncService
   end
 
   # helper method that returns whether an author is up to date with regards to its github issue
-  def author_is_up_to_date(author)
-    github_issue = github_issue_for author
+  def author_is_up_to_date(author, github_issue)
     author.name == github_issue["title"] && author.bio == github_issue["body"]
   end
 
@@ -68,11 +96,10 @@ class GithubSyncService
   end
 
   # This method updates an author in the database from the github issue information
-  def update_author_from_github(author)
+  def update_author_from_github(author, github_issue)
     Rails.logger.info "Updating author #{author.id} from changed GitHub information"
-    github_issue = github_issue_for author
     unless author.update(name: github_issue["title"], bio: github_issue["body"], gh_last_sync: Time.now)
-      Rails.logger.warn "There was an error updating #{author.id} from changed GitHub informatin"
+      Rails.logger.warn "There was an error updating #{author.id} from changed GitHub information"
     end
   end
 
@@ -81,10 +108,19 @@ class GithubSyncService
   def delete_authors_synced_with_no_issue
     authors_to_delete = authors_synced.select { |author| issue_numbers.exclude?(author.gh_issue_number) }
     authors_to_delete.each do |author|
+      delete_author(author, author.gh_issue_number)
+    end
+  end
+
+  # Deletes an author, optionally receives an issue number to emit an error message in case the author passed is nil
+  def delete_author(author, issue_number = nil)
+    if author
       Rails.logger.info "Deleting author #{author.id} from database because issue #{author.gh_issue_number} is no longer in the repo"
       unless author.destroy
         Rails.logger.warn "There was an error deleting #{author.id}"
       end
+    else
+      Rails.logger.warn "Trying to delete an author for issue ##{issue_number} that doesn't exist in the database. Ignoring"
     end
   end
 
@@ -129,6 +165,26 @@ class GithubSyncService
     book = Book.new(title: "#{author.name}'s Opus", author: author, publisher: author, price: 0.99)
     unless book.save
       Rails.logger.warn "There was an error creating a book for author #{author.id} from GitHub issue ##{issue["number"]}"
+    end
+  end
+
+  # Updates an author from a github issue, or creates it with a warning if it doesn't exist
+  def update_or_create_author_from_github(author, github_issue)
+    if author
+      update_author_from_github(author, github_issue)
+    else
+      Rails.logger.warn "Author for edited issue #{github_issue["number"]} doesn't exist. Creating it"
+      create_author_for github_issue
+    end
+  end
+
+  # Creates an author from a github issue, or updates it with a warning if it exists
+  def create_or_update_author_from_github(author, github_issue)
+    if author
+      Rails.logger.warn "Opened issue has an existing author in the DB. How is this possible?. Updating it"
+      update_author_from_github(author, github_issue)
+    else
+      create_author_for github_issue
     end
   end
 end
